move schema tasks creation into uploader class
